requirements.txt
python-dotenv
numpy
pandas
pymongo
certifi
pymongo[srv]==3.11
dill
pyyaml
fastapi
uvicorn

# ML and preprocessing
scikit-learn

# Flask app (for UI)
Flask

# URL feature extraction dependencies
requests
beautifulsoup4
python-whois

# Optional but recommended
joblib
